outbound: Move HTTP balancer into its own module
No functional changes. Setup for changes to come.
