+++ linkerd/app/core/src/config.rs
@@ -30,7 +30,7 @@ pub struct ProxyConfig {
     pub detect_protocol_timeout: Duration,
 }
 
-#[derive(Debug, Copy, Clone)]
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub struct QueueConfig {
     /// The number of requests (or connections, depending on the context) that
     /// may be buffered
