+++ linkerd/app/outbound/src/http/concrete.rs
@@ -1,10 +1,10 @@
 //! A stack that (optionally) resolves a service to a set of endpoint replicas
 //! and distributes HTTP requests among them.
 
-use super::{balance, breaker, client, handle_proxy_error_headers};
+use super::{balance::EwmaConfig, client, handle_proxy_error_headers};
 use crate::{http, stack_labels, BackendRef, Outbound, ParentRef};
 use linkerd_app_core::{
-    classify,
+    config::QueueConfig,
     metrics::{prefix_labels, EndpointLabels, OutboundEndpointLabels},
     profiles,
     proxy::{
@@ -20,6 +20,7 @@ use linkerd_proxy_client_policy::FailureAccrual;
 use std::{fmt::Debug, net::SocketAddr, sync::Arc};
 use tracing::info_span;
 
+mod balance;
 mod metrics;
 #[cfg(test)]
 mod tests;
@@ -29,7 +30,7 @@ pub use self::metrics::BalancerMetrics;
 /// Parameter configuring dispatcher behavior.
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
 pub enum Dispatch {
-    Balance(NameAddr, balance::EwmaConfig),
+    Balance(NameAddr, EwmaConfig),
     Forward(Remote<ServerAddr>, Metadata),
     Fail { message: Arc<str> },
 }
@@ -39,15 +40,6 @@ pub enum Dispatch {
 #[error("{0}")]
 pub struct DispatcherFailed(Arc<str>);
 
-/// Wraps errors encountered in this module.
-#[derive(Debug, thiserror::Error)]
-#[error("{}: {source}", backend.0)]
-pub struct BalanceError {
-    backend: BackendRef,
-    #[source]
-    source: Error,
-}
-
 /// Inner stack target type.
 #[derive(Clone, Debug, PartialEq, Eq)]
 pub struct Endpoint<T> {
@@ -55,15 +47,8 @@ pub struct Endpoint<T> {
     is_local: bool,
     metadata: Metadata,
     parent: T,
+    queue: QueueConfig,
     close_server_connection_on_remote_proxy_error: bool,
-}
-
-/// A target configuring a load balancer stack.
-#[derive(Clone, Debug, PartialEq, Eq)]
-struct Balance<T> {
-    addr: NameAddr,
-    ewma: balance::EwmaConfig,
-    parent: T,
 }
 
 // === impl Outbound ===
@@ -99,6 +84,9 @@ impl<N> Outbound<N> {
         self.map_stack(|config, rt, inner| {
             let inbound_ips = config.inbound_ips.clone();
 
+            // TODO(ver) Configure this from discovery.
+            let queue = config.http_request_queue;
+
             let forward = inner
                 .clone()
                 .push_on_service(
@@ -107,26 +95,27 @@ impl<N> Outbound<N> {
                         .stack
                         .layer(stack_labels("http", "forward")),
                 )
-                // TODO(ver) Configure this queue from the target (i.e. from
-                // discovery).
-                .push(svc::NewQueue::layer_via(config.http_request_queue))
+                .push(svc::NewQueue::layer())
                 .instrument(|e: &Endpoint<T>| info_span!("forward", addr = %e.addr));
 
             let fail = svc::ArcNewService::new(|message: Arc<str>| {
-                svc::mk(move |_| {
-                    let error = DispatcherFailed(message.clone());
-                    futures::future::ready(Err(error))
-                })
+                svc::mk(move |_| futures::future::ready(Err(DispatcherFailed(message.clone()))))
             });
 
             inner
-                .push(Balance::layer(config, rt, resolve))
+                .push(balance::Balance::layer(config, rt, resolve))
+                .check_new_clone()
                 .push_switch(Ok::<_, Infallible>, forward.into_inner())
                 .push_switch(
                     move |parent: T| -> Result<_, Infallible> {
                         Ok(match parent.param() {
                             Dispatch::Balance(addr, ewma) => {
-                                svc::Either::A(svc::Either::A(Balance { addr, ewma, parent }))
+                                svc::Either::A(svc::Either::A(balance::Balance {
+                                    addr,
+                                    ewma,
+                                    parent,
+                                    queue,
+                                }))
                             }
                             Dispatch::Forward(addr, metadata) => svc::Either::A(svc::Either::B({
                                 let is_local = inbound_ips.contains(&addr.ip());
@@ -135,136 +124,17 @@ impl<N> Outbound<N> {
                                     addr,
                                     metadata,
                                     parent,
+                                    queue,
                                     close_server_connection_on_remote_proxy_error: true,
                                 }
                             })),
                             Dispatch::Fail { message } => svc::Either::B(message),
                         })
                     },
-                    fail,
+                    svc::stack(fail).check_new_clone().into_inner(),
                 )
                 .arc_new_clone_http()
         })
-    }
-}
-
-// === impl Balance ===
-
-impl<T> svc::Param<http::balance::EwmaConfig> for Balance<T> {
-    fn param(&self) -> http::balance::EwmaConfig {
-        self.ewma
-    }
-}
-
-impl<T> Balance<T>
-where
-    // Parent target.
-    T: svc::Param<ParentRef>,
-    T: svc::Param<BackendRef>,
-    T: svc::Param<FailureAccrual>,
-    T: Clone + Debug + Send + Sync + 'static,
-{
-    pub fn layer<N, NSvc, R>(
-        config: &crate::Config,
-        rt: &crate::Runtime,
-        resolve: R,
-    ) -> impl svc::Layer<N, Service = svc::ArcNewCloneHttp<Self>> + Clone
-    where
-        // Endpoint resolution.
-        R: Resolve<ConcreteAddr, Error = Error, Endpoint = Metadata> + 'static,
-        // Endpoint stack.
-        N: svc::NewService<Endpoint<T>, Service = NSvc> + Clone + Send + Sync + 'static,
-        NSvc: svc::Service<http::Request<http::BoxBody>, Response = http::Response<http::BoxBody>>
-            + Send
-            + 'static,
-        NSvc::Error: Into<Error>,
-        NSvc::Future: Send,
-    {
-        // TODO(ver) Configure queues from the target (i.e. from discovery).
-        let http_queue = config.http_request_queue;
-        let inbound_ips = config.inbound_ips.clone();
-        let metrics = rt.metrics.clone();
-
-        let resolve = svc::stack(resolve.into_service())
-            .push_map_target(|t: Self| ConcreteAddr(t.addr))
-            .into_inner();
-
-        svc::layer::mk(move |inner: N| {
-            let endpoint = svc::stack(inner)
-                .push_map_target({
-                    let inbound_ips = inbound_ips.clone();
-                    move |((addr, metadata), target): ((SocketAddr, Metadata), Self)| {
-                        tracing::trace!(%addr, ?metadata, ?target, "Resolved endpoint");
-                        let is_local = inbound_ips.contains(&addr.ip());
-                        Endpoint {
-                            addr: Remote(ServerAddr(addr)),
-                            metadata,
-                            is_local,
-                            parent: target.parent,
-                            // We don't close server-side connections when we
-                            // get `l5d-proxy-connection: close` response headers
-                            // going through the balancer.
-                            close_server_connection_on_remote_proxy_error: false,
-                        }
-                    }
-                })
-                .push_on_service(svc::MapErr::layer_boxed())
-                .lift_new_with_target()
-                .push(
-                    http::NewClassifyGateSet::<classify::Response, _, _, _>::layer_via({
-                        let channel_capacity = http_queue.capacity;
-                        move |target: &Self| breaker::Params {
-                            accrual: target.parent.param(),
-                            channel_capacity,
-                        }
-                    }),
-                )
-                .push(balance::NewGaugeEndpoints::layer_via({
-                    let metrics = metrics.http_balancer.clone();
-                    move |target: &Self| {
-                        metrics.http_endpoints(target.parent.param(), target.parent.param())
-                    }
-                }))
-                .push_on_service(svc::OnServiceLayer::new(
-                    metrics.proxy.stack.layer(stack_labels("http", "endpoint")),
-                ))
-                .push_on_service(svc::NewInstrumentLayer::new(
-                    |(addr, _): &(SocketAddr, _)| info_span!("endpoint", %addr),
-                ))
-                .push_on_service(svc::OnServiceLayer::new(svc::BoxService::layer()))
-                .push_on_service(svc::ArcNewService::layer())
-                .push(svc::ArcNewService::layer());
-
-            endpoint
-                .push(http::NewBalancePeakEwma::layer(resolve.clone()))
-                .push(svc::NewMapErr::layer_from_target::<BalanceError, _>())
-                .push_on_service(http::BoxResponse::layer())
-                .push(svc::NewQueue::layer_via(http_queue))
-                .push_on_service(metrics.proxy.stack.layer(stack_labels("http", "balance")))
-                .instrument(|t: &Self| {
-                    let BackendRef(meta) = t.parent.param();
-                    info_span!(
-                        "service",
-                        ns = %meta.namespace(),
-                        name = %meta.name(),
-                        port = %meta.port().map(u16::from).unwrap_or(0),
-                    )
-                })
-                .arc_new_clone_http()
-                .into_inner()
-        })
-    }
-}
-
-// === impl BalanceError ===
-
-impl<T> From<(&Balance<T>, Error)> for BalanceError
-where
-    T: svc::Param<BackendRef>,
-{
-    fn from((target, source): (&Balance<T>, Error)) -> Self {
-        let backend = target.parent.param();
-        Self { backend, source }
     }
 }
 
@@ -273,6 +143,18 @@ where
 impl<T> svc::Param<Remote<ServerAddr>> for Endpoint<T> {
     fn param(&self) -> Remote<ServerAddr> {
         self.addr
+    }
+}
+
+impl<T> svc::Param<svc::queue::Capacity> for Endpoint<T> {
+    fn param(&self) -> svc::queue::Capacity {
+        svc::queue::Capacity(self.queue.capacity)
+    }
+}
+
+impl<T> svc::Param<svc::queue::Timeout> for Endpoint<T> {
+    fn param(&self) -> svc::queue::Timeout {
+        svc::queue::Timeout(self.queue.failfast_timeout)
     }
 }
